+++ NEWS.md
@@ -1,3 +1,6 @@
+# spant 2.18.0
+* Improved automated file type detection.
+
 # spant 2.17.0
 * Simulation duration output is now rounded to 2dp.
 * sim_basis now accepts a character vector of molecular names.